Add 'sponge.command.help' permission to '/sponge:help' command
Resolves #1476
