+++ src/main/java/org/spongepowered/common/SpongeBootstrap.java
@@ -39,7 +39,6 @@ import org.spongepowered.api.service.user.UserStorageService;
 import org.spongepowered.api.service.whitelist.WhitelistService;
 import org.spongepowered.api.util.annotation.NonnullByDefault;
 import org.spongepowered.common.command.SpongeCommand;
-import org.spongepowered.common.command.SpongeHelpCommand;
 import org.spongepowered.common.service.ban.SpongeBanService;
 import org.spongepowered.common.service.pagination.SpongePaginationService;
 import org.spongepowered.common.service.rcon.MinecraftRconService;
@@ -73,7 +72,7 @@ public final class SpongeBootstrap {
 
     public static void initializeCommands() {
         commandManager.register(SpongeImpl.getPlugin(), SpongeCommand.getCommand(), "sponge", "sp");
-        commandManager.register(SpongeImpl.getPlugin(), SpongeHelpCommand.create(), "help", "?");
+        commandManager.register(SpongeImpl.getPlugin(), SpongeCommand.getHelpCommand(), "help", "?");
         commandManager.register(SpongeImpl.getPlugin(), SpongeCallbackHolder.getInstance().createCommand(), SpongeCallbackHolder.CALLBACK_COMMAND);
     }
